+++ src/test/e2e/uffizzicluster_controller_test.go
@@ -4,29 +4,10 @@ import (
 	"context"
 	"github.com/UffizziCloud/uffizzi-cluster-operator/src/api/v1alpha1"
 	"github.com/UffizziCloud/uffizzi-cluster-operator/src/pkg/constants"
-	"github.com/UffizziCloud/uffizzi-cluster-operator/src/test/util/resources"
 	. "github.com/onsi/ginkgo/v2"
 	v1 "k8s.io/api/core/v1"
 )
 
-type TestDefinition struct {
-	Name string
-	Spec v1alpha1.UffizziClusterSpec
-}
-
-func (td *TestDefinition) ExecLifecycleTest(ctx context.Context) {
-	ns := resources.CreateTestNamespace(td.Name)
-	uc := resources.CreateTestUffizziCluster(td.Name, ns.Name)
-	uc.Spec = td.Spec
-	wrapUffizziClusterLifecycleTest(ctx, ns, uc)
-}
-
-const (
-	timeout        = "2m"
-	pollingTimeout = "100ms"
-)
-
-// basic clusters with no configuration
 var _ = Describe("Basic Vanilla K3S UffizziCluster Lifecycle", func() {
 	BeforeEach(func() {
 		if e2e.IsTainted {
@@ -57,7 +38,23 @@ var _ = Describe("Basic Vanilla K8S UffizziCluster Lifecycle", func() {
 	testUffizziCluster.ExecLifecycleTest(ctx)
 })
 
-// clusters with tainted nodes
+var _ = Describe("Basic K3S UffizziCluster with ETCD Lifecycle", func() {
+	BeforeEach(func() {
+		if e2e.IsTainted {
+			Skip("Skipping test because cluster is tainted")
+		}
+	})
+	ctx := context.Background()
+	testUffizziCluster := TestDefinition{
+		Name: "k3s-etcd-test",
+		Spec: v1alpha1.UffizziClusterSpec{
+			ExternalDatastore: constants.ETCD,
+		},
+	}
+	testUffizziCluster.ExecLifecycleTest(ctx)
+})
+
+// Test against cluster with tainted nodes - good for testing node affinities
 
 var _ = Describe("UffizziCluster NodeSelector and Tolerations", func() {
 	BeforeEach(func() {
@@ -80,19 +77,6 @@ var _ = Describe("UffizziCluster NodeSelector and Tolerations", func() {
 					Effect:   "NoSchedule",
 				},
 			},
-		},
-	}
-	testUffizziCluster.ExecLifecycleTest(ctx)
-})
-
-// TODO: k3s with etcd
-
-var _ = Describe("Basic K3S UffizziCluster with ETCD Lifecycle", func() {
-	ctx := context.Background()
-	testUffizziCluster := TestDefinition{
-		Name: "k3s-etcd-test",
-		Spec: v1alpha1.UffizziClusterSpec{
-			ExternalDatastore: constants.ETCD,
 		},
 	}
 	testUffizziCluster.ExecLifecycleTest(ctx)
